Move More Code to Static Runtime
