+++ src/interpreter/ExpressionFragmentizer.cpp
@@ -57,17 +57,6 @@ void ExpressionFragmentizer::fragmentizeBinary()
    }
 }
 
-void ExpressionFragmentizer::fragmentizeHashes()
-{
-   for (auto& type_in : types) {
-      auto& [name, pipe] = pipes.emplace_back();
-      auto& iu_in = generated_ius.emplace_back(type_in, "");
-      auto& iu_out = generated_ius.emplace_back(IR::UnsignedInt::build(8), "");
-      auto& op = pipe.attachSuboperator(ExpressionSubop::build(nullptr, {&iu_out}, {&iu_in}, Type::Hash));
-      name = op.id();
-   }
-}
-
 void ExpressionFragmentizer::fragmentizeFunctions()
 {
    {
@@ -86,7 +75,6 @@ ExpressionFragmentizer::ExpressionFragmentizer()
 {
    fragmentizeBinary();
    fragmentizeCasts();
-   fragmentizeHashes();
    fragmentizeFunctions();
 }
